vdk-control-cli: Move to vdk-core JobConfig
So far, there have been code duplication between vdk-core and vdk-control-cli, with
both components providing a JobConfig class that reads the config.ini file of data jobs
and parses its content. This code duplication creates unnecessary work to keep the logic
in both classes in sync.

This change updates the imports in vdk-control-cli to use the JobConfig class of vdk-core.
In a separate commit, the JobConfig in vdk-control-cli will be deleted to remove the code
duplication.

Testing done: CI/CD

Signed-off-by: Andon Andonov <andonova@vmware.com>